Added light mode and dark mode maps
Also added dark mode and about buttons in their own component, so they can be used on all pages.
