+++ src/Landing.css
@@ -481,7 +481,7 @@ body.dark-mode text{
 }
 
 .full-height-map{
-    @apply h-[30rem] w-[34.5rem] bg-gray-500 flex justify-center items-center mb-[-50px]
+    @apply h-[30rem] w-[34.5rem] bg-gray-500 flex justify-center items-center mb-[-50px] 
 }
 
 .back-to-results{
